Skin Cancer bioinformatics project/requirements.txt
biopython

pandas

matplotlib

seaborn

scikit-bio